Source Code
wiz_test.py is the source for the main program. preview_pattern.py and preview_editor.py is the source for the pattern editor. volume_config_gui.py is the source for the volume visualizer GUI. wiz_visualizer.cpp is the source for the C++ volume visualizer program. config_gui.py is the source for the frequency GUI. wiz_visualizer_freq.cpp is the C++ source for the frequency based visualizer program.